move Signature to proof crate
