--- .samples/didion.json
@@ -1,75 +1,0 @@
-{
-    "did": "did:ion:EiBbvWpBHPql_2W9sObr4vPFl9k9InEDYOEDXvPPRDRyBg",
-    "registrationStatus": "PROCESSING",
-    "localMetadata": {
-        "keys": [
-            {
-                "kmsKeyId": "a21fc9ff-1d67-4d76-a02c-d4524acc1685",
-                "didDocumentKeyId": "did:ion:EiBbvWpBHPql_2W9sObr4vPFl9k9InEDYOEDXvPPRDRyBg#z6MkvvCxe4mL1ghEEgKL"
-            },
-            {
-                "kmsKeyId": "fa4672d1-e833-42e6-8a55-a0da956c2c4c",
-                "didDocumentKeyId": "did:ion:EiBbvWpBHPql_2W9sObr4vPFl9k9InEDYOEDXvPPRDRyBg#z6LSnqJCx9BnhEShY1uC"
-            },
-            {
-                "kmsKeyId": "63b1525e-20a7-4026-b323-a114a66b58e7",
-                "authorizedDidOperations": [
-                    "Recover"
-                ]
-            },
-            {
-                "kmsKeyId": "92f6f4f4-5e46-41f4-a336-88fad589b6b5",
-                "authorizedDidOperations": [
-                    "Update"
-                ]
-            }
-        ],
-        "registered": 1628811763535,
-        "initialDidDocument": {
-            "id": "did:ion:EiBbvWpBHPql_2W9sObr4vPFl9k9InEDYOEDXvPPRDRyBg",
-            "@context": [
-                "https://www.w3.org/ns/did/v1",
-                {
-                    "@base": "did:ion:EiBbvWpBHPql_2W9sObr4vPFl9k9InEDYOEDXvPPRDRyBg"
-                }
-            ],
-            "keyAgreement": [
-                "#z6LSnqJCx9BnhEShY1uC"
-            ],
-            "authentication": [
-                "#z6MkvvCxe4mL1ghEEgKL"
-            ],
-            "assertionMethod": [
-                "#z6MkvvCxe4mL1ghEEgKL"
-            ],
-            "verificationMethod": [
-                {
-                    "id": "#z6MkvvCxe4mL1ghEEgKL",
-                    "type": "JsonWebKey2020",
-                    "controller": "",
-                    "publicKeyJwk": {
-                        "x": "9KIvHIY9zI9TZtrpdx4AIxfsZArbLDKhs3-PBlNLX34",
-                        "crv": "Ed25519",
-                        "kty": "OKP"
-                    }
-                },
-                {
-                    "id": "#z6LSnqJCx9BnhEShY1uC",
-                    "type": "JsonWebKey2020",
-                    "controller": "",
-                    "publicKeyJwk": {
-                        "x": "pca-6GiWWYRSys1U2j5I6gqq7olyJw0M3aqZcuHsST0",
-                        "crv": "X25519",
-                        "kty": "OKP"
-                    }
-                }
-            ],
-            "capabilityDelegation": [
-                "#z6MkvvCxe4mL1ghEEgKL"
-            ],
-            "capabilityInvocation": [
-                "#z6MkvvCxe4mL1ghEEgKL"
-            ]
-        }
-    }
-}
