--- crates/did/src/web/did-eddsa.json
@@ -1,35 +1,0 @@
-{
-    "@context": [
-        "https://www.w3.org/ns/did/v1",
-        "https://w3id.org/security/data-integrity/v1"
-    ],
-    "id": "did:web:demo.credibil.io",
-    "verificationMethod": [
-        {
-            "id": "did:web:demo.credibil.io#key-0",
-            "type": "Multikey",
-            "controller": "did:web:demo.credibil.io",
-            "publicKeyMultibase": "z6Mkr1NtupNezZtcUAMxJ79HPex6ZTR9RnGh8xfV257ZQdss"
-        }
-    ],
-    "authentication": [
-        "did:web:demo.credibil.io#key-0"
-    ],
-    "assertionMethod": [
-        "did:web:demo.credibil.io#key-0"
-    ],
-    "capabilityInvocation": [
-        "did:web:demo.credibil.io#key-0"
-    ],
-    "capabilityDelegation": [
-        "did:web:demo.credibil.io#key-0"
-    ],
-    "keyAgreement": [
-        {
-            "id": "did:web:demo.credibil.io#key-1",
-            "type": "Multikey",
-            "controller": "did:web:demo.credibil.io",
-            "publicKeyMultibase": "z6LSo1jPr1g7qoR9UsYWoJs66FiaoHGAK5cJzBGSvQpuVEQv"
-        }
-    ]
-}
